+++ app/src/main/java/com/picpay/desafio/android/framework/network/PicPayService.kt
@@ -1,11 +1,10 @@
-package com.picpay.desafio.android.network
+package com.picpay.desafio.android.framework.network
 
 import com.picpay.desafio.android.domain.model.User
 import retrofit2.Call
 import retrofit2.http.GET
 
 interface PicPayService {
-
     @GET("users")
     fun getUsers(): Call<List<User>>
 }
