+++ app/src/main/java/com/picpay/desafio/android/presentation/MainActivity.kt
@@ -10,7 +10,7 @@ import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.picpay.desafio.android.R
 import com.picpay.desafio.android.domain.model.User
-import com.picpay.desafio.android.network.PicPayService
+import com.picpay.desafio.android.framework.network.PicPayService
 import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
