+++ app/src/test/java/com/picpay/desafio/android/ExampleService.kt
@@ -1,7 +1,7 @@
 package com.picpay.desafio.android
 
 import com.picpay.desafio.android.domain.model.User
-import com.picpay.desafio.android.network.PicPayService
+import com.picpay.desafio.android.framework.network.PicPayService
 
 class ExampleService(
     private val service: PicPayService
