+++ app/src/test/java/com/picpay/desafio/android/ExampleServiceTest.kt
@@ -3,7 +3,7 @@ package com.picpay.desafio.android
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import com.picpay.desafio.android.domain.model.User
-import com.picpay.desafio.android.network.PicPayService
+import com.picpay.desafio.android.framework.network.PicPayService
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import retrofit2.Call
